requirements.txt
langchain~=0.3.21
faiss-cpu~=1.10.0
openai~=1.69.0
python-dotenv
beautifulsoup4~=4.13.3
requests~=2.32.3
streamlit~=1.44.0
tiktoken~=0.9.0
playwright~=1.51.0
soupsieve~=2.6
packaging~=24.2
GitPython~=3.1.44
gitdb~=4.0.12
h11~=0.14.0
numpy~=2.0.2
future~=0.18.2
pillow~=11.1.0
pip~=21.2.4
wheel~=0.37.0
protobuf~=5.29.4
tornado~=6.4.2
pytz~=2025.2
toml~=0.10.2
PyYAML~=6.0.2
exceptiongroup~=1.2.2
sniffio~=1.3.1
attrs~=25.3.0
setuptools~=58.0.4
smmap~=5.0.2
click~=8.1.8
httpcore~=1.0.7
idna~=3.10
httpx~=0.28.1
zstandard~=0.23.0
jiter~=0.9.0
charset-normalizer~=3.4.1
regex~=2024.11.6
altair~=5.5.0
jsonschema~=4.23.0
narwhals~=1.32.0
pandas~=2.2.3
pyarrow~=19.0.1
jinja2~=3.1.6
referencing~=0.36.2
distro~=1.9.0
MarkupSafe~=3.0.2
pydantic~=2.11.1
anyio~=4.9.0
orjson~=3.10.16
sqlalchemy~=2.0.40
python-dateutil~=2.9.0.post0
pydeck~=0.9.1
certifi~=2025.1.31
tzdata~=2025.2
urllib3~=2.3.0
six~=1.15.0
annotated-types~=0.7.0
tenacity~=9.0.0
langchain-core~=0.3.49
langsmith~=0.3.19
langchain-text-splitters~=0.3.7
async-timeout~=4.0.3
requests-toolbelt~=1.0.0
cachetools~=5.5.2
blinker~=1.9.0
jsonpointer~=3.0.0
jsonschema-specifications~=2024.10.1
greenlet~=3.1.1
jsonpatch~=1.33